Extract new validation tools to the corresponding module
